# contracts/openapi/ESG_P0_OpenAPI_v0.1.yaml
openapi: 3.0.3
info:
  title: ESG报告软件 P0 API
  version: 0.1.0
  description: ESG报告软件MVP阶段P0接口规格，覆盖多租户、项目、同行报告解析、议题推荐、采集审核、知识库、AI写作校对、导出、AI成本与审计。
servers:
- url: /api/v1
  description: 默认API前缀
...
paths:
...
  /enterprises:
    post:
      tags:
      - 企业管理
      summary: 创建企业
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enterprise'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnterpriseCreateRequest'
      security:
      - bearerAuth: []
    get:
      tags:
      - 企业管理
      summary: 查询企业列表
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnterpriseListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - name: keyword
        in: query
        required: false
        schema:
          type: string
      - name: exchange
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
      security:
      - bearerAuth: []
  /enterprises/{enterprise_id}:
    parameters:
    - name: enterprise_id
      in: path
      required: true
      schema:
        type: string
    get:
      tags:
      - 企业管理
      summary: 获取企业详情
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enterprise'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
    patch:
      tags:
      - 企业管理
      summary: 更新企业
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enterprise'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnterpriseCreateRequest'
      security:
      - bearerAuth: []
...
  /projects:
    post:
      tags:
      - 报告项目
      summary: 创建报告项目
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateRequest'
      security:
      - bearerAuth: []
    get:
      tags:
      - 报告项目
      summary: 查询项目列表
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      parameters:
      - name: enterprise_id
        in: query
        required: false
        schema:
          type: string
      - name: report_year
        in: query
        required: false
        schema:
          type: integer
      - name: project_status
        in: query
        required: false
        schema:
          type: string
      - name: keyword
        in: query
        required: false
        schema:
          type: string
      security:
      - bearerAuth: []
  /projects/{project_id}/dashboard:
    parameters:
    - name: project_id
      in: path
      required: true
      schema:
        type: string
    get:
      tags:
      - 报告项目
      summary: 项目工作台
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDashboard'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearerAuth: []
  /projects/{project_id}:
    parameters:
    - name: project_id
      in: path
      required: true
      schema:
        type: string
    get:
      tags:
      - 报告项目
      summary: 获取项目详情
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
      security:
      - bearerAuth: []
    patch:
      tags:
      - 报告项目
      summary: 更新项目
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
      security:
      - bearerAuth: []
  /projects/{project_id}/status:
    parameters:
    - name: project_id
      in: path
      required: true
      schema:
        type: string
    post:
      tags:
      - 报告项目
      summary: 项目状态流转
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectStatusTransitionRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
      security:
      - bearerAuth: []
  /projects/{project_id}/members:
    parameters:
    - name: project_id
      in: path
      required: true
      schema:
        type: string
    get:
      tags:
      - 报告项目
      summary: 查询项目成员
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMemberListResponse'
      security:
      - bearerAuth: []
    post:
      tags:
      - 报告项目
      summary: 添加项目成员
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectMemberCreateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectMemberListResponse'
      security:
      - bearerAuth: []
  /projects/{project_id}/members/{user_id}:
    parameters:
    - name: project_id
      in: path
      required: true
      schema:
        type: string
    - name: user_id
      in: path
      required: true
      schema:
        type: string
    delete:
      tags:
      - 报告项目
      summary: 移除项目成员
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      required:
      - success
      - data
      - message
      - request_id
      - timestamp
      properties:
        success:
          type: boolean
        data:
          nullable: true
        message:
          type: string
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
      - success
      - message
      - error
      properties:
        success:
          type: boolean
          example: false
        data:
          nullable: true
        message:
          type: string
        error:
          $ref: '#/components/schemas/ErrorObject'
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
    ErrorObject:
      type: object
      properties:
        code:
          type: string
        type:
          type: string
        details:
          type: object
          additionalProperties: true
    PagedResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    EnterpriseCreateRequest:
      type: object
      required:
      - enterprise_name
      properties:
        enterprise_name:
          type: string
        enterprise_short_name:
          type: string
        stock_code:
          type: string
        exchange:
          type: string
        country_or_region:
          type: string
        industry_description:
          type: string
        main_business:
          type: string
        status:
          type: string
          enum:
          - active
          - inactive
    Enterprise:
      type: object
      properties:
        enterprise_id:
          type: string
        tenant_id:
          type: string
        enterprise_name:
          type: string
        enterprise_short_name:
          type: string
        stock_code:
          type: string
        exchange:
          type: string
        country_or_region:
          type: string
        industry_description:
          type: string
        main_business:
          type: string
        gics:
          $ref: '#/components/schemas/GicsClassification'
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EnterpriseListResponse:
      allOf:
      - $ref: '#/components/schemas/PagedResponse'
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Enterprise'
    GicsClassification:
      type: object
      properties:
        gics_level_1_code:
          type: string
        gics_level_1_name:
          type: string
        gics_level_2_code:
          type: string
        gics_level_2_name:
          type: string
        gics_level_3_code:
          type: string
        gics_level_3_name:
          type: string
        gics_level_4_code:
          type: string
        gics_level_4_name:
          type: string
...
    ProjectCreateRequest:
      type: object
      required:
      - enterprise_id
      - project_name
      - report_year
      - report_type
      - project_owner_user_id
      properties:
        enterprise_id:
          type: string
        project_name:
          type: string
        report_year:
          type: integer
        report_type:
          type: string
        report_language:
          type: string
          enum:
          - zh
          - en
          - bilingual
        reporting_period_start:
          type: string
          format: date
        reporting_period_end:
          type: string
          format: date
        report_boundary:
          type: string
        selected_standard_codes:
          type: array
          items:
            type: string
        project_owner_user_id:
          type: string
        reuse_from_project_id:
          type: string
          nullable: true
        reuse_options:
          type: object
          additionalProperties:
            type: boolean
    ProjectUpdateRequest:
      type: object
      properties:
        project_name:
          type: string
        report_year:
          type: integer
        report_type:
          type: string
        report_language:
          type: string
          enum:
          - zh
          - en
          - bilingual
        reporting_period_start:
          type: string
          format: date
        reporting_period_end:
          type: string
          format: date
        report_boundary:
          type: string
        selected_standard_codes:
          type: array
          items:
            type: string
        project_owner_user_id:
          type: string
    ProjectStatusTransitionRequest:
      type: object
      required:
      - target_status
      properties:
        target_status:
          type: string
    ProjectMemberCreateRequest:
      type: object
      required:
      - user_id
      - project_role
      properties:
        user_id:
          type: string
        project_role:
          type: string
        org_unit_id:
          type: string
          nullable: true
    ProjectMember:
      type: object
      properties:
        project_member_id:
          type: string
        project_id:
          type: string
        user_id:
          type: string
        name:
          type: string
        email:
          type: string
        project_role:
          type: string
        org_unit_id:
          type: string
          nullable: true
        status:
          type: string
    ProjectMemberListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProjectMember'
    Project:
      type: object
      properties:
        project_id:
          type: string
        enterprise_id:
          type: string
        project_name:
          type: string
        report_year:
          type: integer
        project_status:
          type: string
          enum:
          - draft
          - peer_analysis
          - topic_confirmation
          - task_assignment
          - data_collection
          - department_review
          - writing
          - chapter_review
          - full_review
          - export_ready
          - completed
          - archived
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProjectListResponse:
      allOf:
      - $ref: '#/components/schemas/PagedResponse'
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Project'
    ProjectDashboard:
      type: object
      properties:
        project:
          $ref: '#/components/schemas/Project'
        progress:
          type: object
          additionalProperties:
            type: integer
        risks:
          type: array
          items:
            $ref: '#/components/schemas/RiskItem'
        next_steps:
          type: array
          items:
            type: object
            properties:
              step_code:
                type: string
              message:
                type: string
              target_url:
                type: string
    RiskItem:
      type: object
      properties:
        risk_id:
          type: string
        risk_type:
          type: string
        severity:
          type: string
          enum:
          - high
          - medium
          - low
        message:
          type: string
        target_url:
          type: string
...